Remove GroupConfigManager validate
